builders: use yocto-autobuilder-helper's clobberdir script
Removing the build directory (with rm -fr) can take a significant amount
of time, longer than the standard buildbot timeout, which delays the
builds. Instead, we use the clobberdir script from the helper repository.

This script moves the directory contents out of the way to a location which
is reaped by a background process.
Unfortunately this assumes/requires a copy of the repository is kept in the
home directory of the user running buildbot.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>