builders: implement get_publish_dest()
Based on the SetDest step in the old codebase this function is used to
determine where the artefacts of a build should be published. In order to
avoid overwriting previously published content the function keeps a list of
destinations which have already been created and written to on the
controller and avoids reusing them.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>