builders: fix get_publish_dest()
Various errors resulted from not having been able to test the initial
implmentation of get_publish_dest():
* incorrect logic to determine whether a build was a release
* passing a filename rather than a string representing file contents to
  json.loads()
* missing parameter on a call to buildbot's IProperties.setProperty()

Further we update the properties passed to the triggered builds via the
Trigger build steps to include the deploy_artefacts and publish_destination
properties.

This patch is able to succesfully supply required information for the
helper scripts to publish artefacts.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>