wikilog: Complete porting to new buildbot codebase and py3
Finish the porting work started by Joshua Lock, accounting for changes
in buildbot APIs/data model and changes from py3, particular around
character encoding.

This also changes the behaviour of the plugin slightly. We now
use the build URL in the header to match builds. With the new codebase
we can walk the parent tree of triggers builds to ensure we always have
the correct parent build url. This means we can drop a lot of the older
more imprecise build matching logic.

Also simplify the format in the wiki log to one output format which lists
all step failures for each build, even in the parent case.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>