wikilog: Ensure we use a separate thread for blocking wiki network access
Buildbot would hang for several minutes when cancelling builds. We need to defer
this work (which involved network calls) to a thread so that the reactor isn't
blocked.

Use a lock to ensure we only update the wiki one entry at a time.

Also tweak the logging to allow easier debugging (include the parent build ID)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>