wikilog: Filter monitoring on wiki log to build subset
We want the wiki log to primarily contain full and quick builds for
processing by SWAT. The buildperf log entries are distracting but
in case of failure we do need them there so adjust the code to handle
this.

Also, don't log reports of skipped builds but do log build warnings, since
we do want warnings to be acted upon now.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>